social-bot/requirements.txt
selenium==4.16.0
webdriver-manager==4.0.1
google-api-python-client==2.111.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.0
schedule==1.2.1
undetected-chromedriver==3.5.5
